+++ tech-motisha/src/Components/Audio/AudioPlayer.css
@@ -1,0 +1,89 @@
+/* Style for the audio player container */
+  .audio-player-container {
+  align-items: center;
+  background-color: #7ca4f3;
+  border: 1px solid #ccc;
+  display: flex;
+  flex-direction: column;
+  padding: 10px;
+  max-width: 400px;
+  margin-bottom: 20px;
+  width: 100%;
+  }
+ 
+  /* Style for the poster */
+  .poster {
+  width: 100%;
+  height: auto;
+  }
+  
+.buttons-container {
+  display: flex;
+  gap: 1em;
+  margin-top: 1em;
+  }
+
+button {
+  background-color: darkblue;
+  border: none;
+  color: #fff;
+  cursor: pointer;
+  font-size: 1em;
+  padding: 0.5em 1em;
+  }
+  
+button:hover {
+  background-color: #fa510f;
+  }
+  
+  /* Style for the comment submit button */
+  .comments-container button[type="submit"] {
+  background-color: darkblue;
+  color: #fa510f;
+  border: none;
+  border-radius: 5px;
+  padding: 10px;
+  cursor: pointer;
+  }
+  
+.subscription-container p {
+  font-size: 1em;
+  font-weight: bold;
+  }
+  
+  /* Style for the subscribe input */
+  .subscribe-container input[type="email"] {
+  border: 3px solid #0a0909;
+  border-radius: 5px;
+  margin-bottom: 10px;
+  padding: 10px;
+  width: 100%;
+  font-size: 1em;
+  padding: 0.5em;
+  }
+  
+.comment-container button[type="submit"],
+.subscription-container button[type="submit"] {
+  background-color: darkblue;
+  border: none;
+  border-radius: 5px;
+  padding: 10px;
+  cursor: pointer;
+  font-size: 1em;
+  padding: 0.5em 1em;
+  }
+
+.comment-container button[type="submit"]:hover,
+.subscription-container button[type="submit"]:hover {
+  background-color: #fa510f;
+}
+
+.comment-container ul {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.comment-container li {
+  margin-bottom: 0.5em;
+}
