+++ tech-motisha/src/Components/assets/music.jsx
@@ -1,0 +1,12 @@
+const musicData = [
+  {
+    id: 1,
+    title: 'The Song',
+    artist: 'Ken',
+    poster: 'https://i.scdn.co/image/ab67616d00001e02f4a2ccbe20d6d52f16816812',
+    duration: 240,
+    source: 'https://www.soundhelix.com/examples/mp3/SoundHelix-Song-2.mp3',
+   },
+];
+
+export default musicData;
