--- tech-motisha/src/assets/music.jsx
@@ -1,22 +1,0 @@
-const musicData = [
-  {
-    id: 1,
-    title: 'The Song',
-    artist: 'Ken',
-    duration: 180,
-    src: 'https://www.soundhelix.com/examples/mp3/SoundHelix-Song-1.mp3',
-    poster: 'https://i.scdn.co/image/ab67616d00001e02f4a2ccbe20d6d52f16816812',
-  },
-  {
-    id: 2,
-    title: 'My man',
-    artist: 'Lucy',
-    duration: 240,
-    src: 'https://www.soundhelix.com/examples/mp3/SoundHelix-Song-2.mp3',
-    poster: 'https://via.placeholder.com/200x200.png?text=Song+2+Poster',
-  },
-
-];
-
-
-export default musicData;
