--- tech-motisha/src/components/AudioPlayer.css
@@ -1,98 +1,0 @@
-.my-header {
-  background-color: #101f3c;
-  color: #fff;
-  padding: 20px;
-}
-/* Style for the audio player container */
-  .audio-player-container {
-  align-items: center;
-  background-color: #7ca4f3;
-  border: 1px solid #ccc;
-  display: flex;
-  flex-direction: column;
-  padding: 10px;
-  max-width: 400px;
-  margin: 20px;
-  width: 100%;
-  }
-  
-  /* Style for the poster container */
-  .poster-container {
-  flex-basis: 30%; /* set width of poster image */
-  margin: 20px; /* add margin to separate from song title */
-  }
-  
-  /* Style for the poster */
-  .poster {
-  width: 100%;
-  height: auto;
-  }
-  
-  /* Style for the song info */
-  .song-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  /* margin-bottom: 10px add margin to separate from audio controls */
-  color: #fa510f;
-  }
-
-  /* Style for the audio control buttons */
-  .audio-control button {
-  margin: 5px;
-  padding: 10px;
-  background-color: #222;
-  color: #fa510f;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  }
-  
-  /* Style for the comment container */
-  .comments-container {
-  margin-top: 10px;
-  }
-  
-  /* Style for the comment input */
-  .comments-container input[type="text"] {
-  width: 100%;
-  padding: 10px;
-  border: 3px solid #1a1313;
-  border-radius: 5px;
-  margin-bottom: 10px;
-  }
-  
-  /* Style for the comment submit button */
-  .comments-container button[type="submit"] {
-  background-color: #222;
-  color: #fa510f;
-  border: none;
-  border-radius: 5px;
-  padding: 10px;
-  cursor: pointer;
-  }
-  
-  /* Style for the subscribe container */
-  .subscribe-container {
-  margin-top: 10px;
-  }
-  
-  /* Style for the subscribe input */
-  .subscribe-container input[type="email"] {
-  border: 3px solid #0a0909;
-  border-radius: 5px;
-  margin-bottom: 10px;
-  padding: 10px;
-  width: 100%;
-  }
-  
-  /* Style for the subscribe button */
-  .subscribe-container button[type="submit"] {
-  background-color: #222;
-  color: #fa510f;
-  border: none;
-  border-radius: 5px;
-  padding: 10px;
-  cursor: pointer;
-  }
